requirements.txt
Django==1.9.1
djangorestframework==3.3.2
gomill==0.7.4
wheel==0.24.0
